feat(eslint): rename sort config to perfectionist
