+++ packages/eslint/src/modules/perfectionist.ts
